electron-test-runner/Test-Foundry/TestFoundry_Framework/requirements.txt
# TestFoundry Framework Dependencies

# Core Framework
pyyaml>=6.0
pandas>=2.0.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# AI Clients (OpenAI package works for both OpenAI and Azure OpenAI)
openai>=1.12.0

# Document Processing - PDF
PyPDF2>=3.0.1
PyMuPDF>=1.23.0
pdfplumber>=0.10.0

# Document Processing - Word
python-docx>=0.8.11

# Document Processing - Excel  
openpyxl>=3.1.0
xlrd>=2.0.1

# Image Processing and OCR
Pillow>=10.0.0
pytesseract>=0.3.10
opencv-python>=4.8.0

# Text Processing
nltk>=3.8.1
chardet>=5.2.0
tiktoken>=0.5.0

# Async and Concurrency
aiohttp>=3.9.0

# Utilities
tqdm>=4.66.0
colorama>=0.4.6
rich>=13.7.0
tenacity>=8.2.0

# Optional: Machine Learning (for advanced text analysis)
scikit-learn>=1.3.0

# Development and Debugging
black>=23.0.0
flake8>=6.0.0